+++ app/libraries/Theme/Setups/Translations/Core.php
@@ -21,7 +21,7 @@ final class Core extends AbstractTranslation
     {
         parent::__construct($theme);
 
-        $this->textDomain = $this->app->get()->get('TextDomain');
+        $this->textDomain = $this->app->theme->get('TextDomain');
     }
 
     public function run()